requirements.txt
google-genai

streamlit 

python-dotenv